Modification de l'interface Viewable
Changement du type de retour de la méthode getWindowPos.

+++ include/contracts/Viewable.hpp
@@ -27,5 +27,5 @@ class Viewable {
     /*
     Retourne la position de la zone graphique du Viewable relativement au coin supérieur gauche de la fenêtre.
     */
-    virtual sf::Vector2f getWindowPos0() = 0;
+    virtual sf::Vector2i getWindowPos0() = 0;
 };
